# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: IMS SEPCUNE API
  description: Comprehensive Inventory Management System API
  version: 1.0.0
  contact:
    name: IMS SEPCUNE Support
    email: support@ims-sepcune.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.ims-sepcune.com
    description: Production server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        status:
          type: string

    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        unitPrice:
          type: number
        categoryId:
          type: string

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new tenant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - subdomain
                - adminEmail
                - adminPassword
                - adminFirstName
                - adminLastName
              properties:
                name:
                  type: string
                subdomain:
                  type: string
                adminEmail:
                  type: string
                adminPassword:
                  type: string
                adminFirstName:
                  type: string
                adminLastName:
                  type: string
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'

    post:
      tags:
        - Users
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User created successfully

  /api/items:
    get:
      tags:
        - Items
      summary: Get all items
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Items retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Item'

    post:
      tags:
        - Items
      summary: Create new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - sku
              properties:
                name:
                  type: string
                sku:
                  type: string
                description:
                  type: string
                unitPrice:
                  type: number
                categoryId:
                  type: string
      responses:
        '201':
          description: Item created successfully

  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

    post:
      tags:
        - Categories
      summary: Create new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Category created successfully

  /api/inventory:
    get:
      tags:
        - Inventory
      summary: Get inventory
      parameters:
        - name: warehouse
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Inventory retrieved successfully

  /api/inventory/receive:
    post:
      tags:
        - Inventory
      summary: Receive stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - itemId
                - warehouseId
                - quantity
                - unitCost
              properties:
                itemId:
                  type: string
                warehouseId:
                  type: string
                quantity:
                  type: number
                unitCost:
                  type: number
                grnNumber:
                  type: string
      responses:
        '201':
          description: Stock received successfully